+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 [unreleased]
 
+#### 1.2.0 / 2021-12-17
+* updated for more parity with planned code
+* updated version check for revert
+* update to use `move_dir()` instead of `$wp_filesystem->move()`
+
 #### 1.1.3 / 2021-09-17
 * update version check